// src/domain/underwriting/du/generated/arcroles.ts
export const DU_RELATIONSHIP_XPATH = "MESSAGE/DEAL_SETS/DEAL_SET/DEALS/DEAL/RELATIONSHIPS/RELATIONSHIP";

export const DU_ARCROLES: Readonly<Record<string, DuArcRole>> = {
// ...
  "UNDERWRITING_VERIFICATION_IsAssociatedWith_ASSET": {
    "arcrole": "urn:fdc:mismo.org:2009:residential/UNDERWRITING_VERIFICATION_IsAssociatedWith_ASSET",
    "name": "UNDERWRITING_VERIFICATION_IsAssociatedWith_ASSET",
    "verbPhrase": "IsAssociatedWith",
    "from": {
      "xpath": "DEAL/LOANS/LOAN/EXTENSION/OTHER/DU:LOAN_EXTENSION/DU:UNDERWRITING_VERIFICATIONS/DU:UNDERWRITING_VERIFICATION",
      "container": "UNDERWRITING_VERIFICATION",
      "relationshipEnd": "UNDERWRITING_VERIFICATION",
      "arcroleTerm": "UNDERWRITING_VERIFICATION",
      "disputed": false
    },
    "to": {
      "xpath": "DEAL/ASSETS/ASSET/OWNED_PROPERTY/OWNED_PROPERTY_DETAIL",
      "container": "OWNED_PROPERTY_DETAIL",
      "relationshipEnd": "ASSET",
      "arcroleTerm": "ASSET",
      "disputed": true
    },
    "note": "If the Borrower's Rental Income amount is to be calculated, an arcrole must be used to associate the corresponding calculation with the ASSET (OWNED_PROPERTY_DETAIL)",
    "exercised": false
  },
  "UNDERWRITING_VERIFICATION_IsAssociatedWith_EMPLOYER": {
    "arcrole": "urn:fdc:mismo.org:2009:residential/UNDERWRITING_VERIFICATION_IsAssociatedWith_EMPLOYER",
    "name": "UNDERWRITING_VERIFICATION_IsAssociatedWith_EMPLOYER",
    "verbPhrase": "IsAssociatedWith",
    "from": {
      "xpath": "DEAL/LOANS/LOAN/EXTENSION/OTHER/DU:LOAN_EXTENSION/DU:UNDERWRITING_VERIFICATIONS/DU:UNDERWRITING_VERIFICATION",
      "container": "UNDERWRITING_VERIFICATION",
      "relationshipEnd": "UNDERWRITING_VERIFICATION",
      "arcroleTerm": "UNDERWRITING_VERIFICATION",
      "disputed": false
    },
    "to": {
      "xpath": "DEAL/PARTIES/PARTY/ROLES/ROLE/BORROWER/EMPLOYERS/EMPLOYER",
      "container": "EMPLOYMENT",
      "relationshipEnd": "EMPLOYMENT",
      "arcroleTerm": "EMPLOYER",
      "disputed": true
    },
    "note": "If the Borrower's Employment Income amount is to be calculated, an arcrole must be used to associate the corresponding calculation with the EMPLOYER (EMPLOYMENT)",
    "exercised": false
  },
// ...
} as const;