// src/domain/underwriting/du/generated/cardinality.ts
export const DU_CARDINALITY: Readonly<Record<string, DuContainerCardinality>> = {
  "MESSAGE": {
    "container": "MESSAGE",
    "du": {
      "min": 1,
      "max": 1
    },
    "fhaVa": {
      "min": 1,
      "max": 1
    }
  },
  "MESSAGE/ABOUT_VERSIONS/ABOUT_VERSION": {
    "container": "ABOUT_VERSION",
    "du": {
      "min": 1,
      "max": 1
    },
    "fhaVa": {
      "min": 1,
      "max": 1
    }
  },
  "MESSAGE/DEAL_SETS/DEAL_SET": {
    "container": "DEAL_SET",
    "du": {
      "min": 1,
      "max": 1
    },
    "fhaVa": {
      "min": 1,
      "max": 1
    }
  },
// ...
} as const;